move didYouMeanQuotedResolver to a separate file
